- Rebased on top of Harry's PR. - Made fixes to config paths - moved files around to make CLI commands clearer
