+++ contracts/sample-apps/SimpleStableCoin.sol
@@ -25,35 +25,3 @@ contract SimpleStableCoin is ERC20 {
         return s_owner;
     }
 }
-
-// TODO @zubin cleanup
-// pragma solidity ^0.8.7;
-
-// import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-// import "@openzeppelin/contracts/token/ERC20/extensions/ERC20Burnable.sol";
-// import "@openzeppelin/contracts/security/Pausable.sol";
-// import "@openzeppelin/contracts/access/Ownable.sol";
-// import "@openzeppelin/contracts/token/ERC20/extensions/draft-ERC20Permit.sol";
-// contract StableCoin is ERC20, ERC20Burnable, Pausable, Ownable, ERC20Permit {
-//     constructor() ERC20("StableCoin", "STC") ERC20Permit("StableCoin") {}
-
-//     function pause() public onlyOwner {
-//         _pause();
-//     }
-
-//     function unpause() public onlyOwner {
-//         _unpause();
-//     }
-
-//     function mint(address to, uint256 amount) public onlyOwner {
-//         _mint(to, amount);
-//     }
-
-//     function _beforeTokenTransfer(address from, address to, uint256 amount)
-//         internal
-//         whenNotPaused
-//         override
-//     {
-//         super._beforeTokenTransfer(from, to, amount);
-//     }
-// }
